+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to Mystic CMake will be documented here.
 
 ### Added
 
+- CMake Module: Moduke Importing Helper
 - CMake Module: Third Party Recursive Inclusion
 - CMake Module: Third Party Dependency Config Helper
 - CMake Module: Coverage Tool